add new wait component apps_running
